Function References

Hyperparameter optimization

LazyGPR.MarginalLikelihood — Type

struct MarginalLikelihood

This is a trait data type that indicates the marginal likelihood should be used.

LazyGPR.LOOCV — Type

struct LOOCV

This is a trait data type that indicates the leave-one-out cross-validation should be used. See Gaussian Process for Machine Learning (Rasmussen, Williams, 2005, DOI: 10.7551/mitpress/3206.001.0001) for more details.

LazyGPR.HCostConfig — Type

struct HCostConfig

This is a trait data type that indicates single-process hyperparameter inferenec should be performed.

LazyGPR.LazyHCostConfig — Type

struct LazyHCostConfig

This is a trait and config data type that indicates multi-process hyperparameter inferenec should be performed. It also specifies that the model is the lazy-evaluation GPR, as opposed to a conventional GPR.

LazyGPR.MetaheuristicsConfig — Type

@kwdef struct MetaheuristicsConfig <: SolverConfig
    f_calls_limit::Integer = 1_000
end

The optimization configuration for Metaheuristics.jl's

LazyGPR.KernelOptimConfig — Type

@kwdef struct KernelOptimConfig{T}
    a_lb::T
    a_ub::T
    κ_ub::T
    N_initials_a::Int = 100
    N_initials_κ::Int = 100
    width_factor::T = one(T)
    height_factor::T = convert(T, 1/2)
end

This configuration is used for the two-stage approach to optimizing DE kernels.

In the first stage, the stationary kernel is optimized.

In the second stage, a canonical kernel is formed based on the solution of the stationary kernel's hyperparameters, and the gain parameter is fitted with the canonical kernel bandwidth being fixed.

About some of the field members:

N_initials_a and N_initials_κ specifies how many initial guesses are used in the optimization algorithm, sampled at uniformly spaced intervals from the lowerbound to the upperbounds.

The gain parameter is κ.

The bandwidth parameter is a for the SqExpKernel and WendlandSpline kernels.

width_factor and height_factor contols how much larger or smaller the DE kernel should be with respect to the bandwidth parameter a from the stationary kernel's optimized solution.

The default setting is for the case when the DE kernel should use the stationary kernel's solution as its canonical kernel.

LazyGPR.UseMetaheuristics — Type

struct UseMetaheuristics

This is a trait data type that indicates Metaheuristics.jl should be used for optimization. Usage:

import Metaheuristics as EVO
opt_trait = UseMetaheuristics(EVO)

LazyGPR.OptimContainer — Type

struct OptimContainer{T}
    lbs::Vector{T}
    ubs::Vector{T}
    p0s::Vector{Vector{T}}
end

A container for optimization problems. lbs and ubs are the lower and upper bounds, respectively. p0s is an array of initial guesses.

LazyGPR.create_hp_cost — Function

create_hp_cost(::UseDistributedComputing, args...)

Creates the hyperparameter cost function for a distributed computing setup.

create_hp_cost(::Nothing, args...)

Creates the hyperparameter cost function for a single-process setup.

LazyGPR.free_hp_cost — Function

free_hp_cost(Bs_info::DData.Dinfo)

Frees up the data specified in Bs_info, which was used by create_hp_cost and run_optimizer.

free_hp_cost(::Nothing, args...)

There is no workers to free for a single-process setup. This function does nothing.

LazyGPR.optimize_kernel_hp — Function

optimize_kernel_hp(
    alg_trait::ExtensionPkgs,
    kernel::PositiveDefiniteKernel,
    ms_trait::ModelSelection,
    model::Union{LazyGP,GPData},
    config::HyperparameterInferenceInfo,
    solver_config::SolverConfig,
    optim_info::OptimContainer{T},
) where T <: Real

Fits the GPR hyperparameters for a single-process setup.

optimize_kernel_hp(
    alg_trait::ExtensionPkgs,
    ref_sk::StationaryKernel,
    ms_trait::ModelSelection,
    model::Union{LazyGP,GPData},
    config::HyperparameterInferenceInfo,
    solver_config::SolverConfig,
    optim_config::KernelOptimConfig,
)

Fits the GPR hyperparameters for multi-process, distributed computing setup.

LazyGPR.optimize_kernel_hp_separately — Function

optimize_kernel_hp_separately(
    alg_trait::ExtensionPkgs,
    ref_dek::DEKernel,
    ms_trait::ModelSelection,
    model::Union{LazyGP,GPData},
    config::HyperparameterInferenceInfo,
    solver_config::SolverConfig,
    optim_config::KernelOptimConfig{T},
) where T <: Real

The two-stage approach to optimizing the DE kernel hyperparameters.

Warp sample construction

LazyGPR.create_grid_warp_samples — Function

create_grid_warp_samples(
    alg_trait::UseSpatialGSP,
    y_nD::AbstractArray{T},
    warp_config;
    smooth_iters::Integer = 0,
) where T

Returns an array of warp samples, constructed using the Bernstein filtering method with a unit grid graph.

In the current implementation, warp_config should be a data type of SpatialGSP.WarpConfig.

As a post-processing step, smooth_iters performs smoothing by applying the adjacency matrix to the warp samples smooth_iters times.

create_grid_warp_samples(
    alg_trait::UseRieszDSP,
    y_nD::AbstractArray{T};
    N_scales = 0
) where T

Returns an array of warp samples, constructed using the higher-order Riesz-wavelet transform, implemented in the RieszDSP.jl package.

N_scales is the number of wavelet bands to use. The default value of 0 means the following is used:

N_scales = round(Int, log2( maximum(size(y_nD))))

LazyGPR.create_warp_samples — Function

function create_warp_samples(
    alg_trait::UseSpatialGSP,
    graph_trait::GraphTrait,
    X::Vector{AV},
    y::Vector{T},
    graph_config,
    warp_config;
    smooth_iters::Integer = 0,
) where {T <: Real, AV <: AbstractVector{T}}

Returns an array of warp samples, constructed using the Bernstein filtering method with a graph construction method specified by graph_trait.

In the current implementation, warp_config should be a data type of SpatialGSP.WarpConfig, and graph_trait can be of the dispatch types AxisGraph and KNNGraph.

As a post-processing step, smooth_iters performs smoothing by applying the adjacency matrix to the warp samples smooth_iters times.

LazyGPR.AxisGraph — Type

struct AxisGraph

This is a trait data type that indicates the axis-search graph construction method should be used.

LazyGPR.KNNGraph — Type

struct KNNGraph
    This is a trait data type that indicates the k-nearest neighbors graph construction method, with forced symmetry to generate undirected graphs, should be used.

LazyGPR.UseRieszDSP — Type

struct UseRieszDSP

This is a trait data type that indicates RieszDSP.jl should be used for constructing the warp samples. Usage:

import RieszDSP as RZ
warp_graph_trait = UseRieszDSP(RZ)

LazyGPR.UseSpatialGSP — Type

struct UseSpatialGSP

This is a trait data type that indicates SpatialGSP.jl should be used for constructing the warp samples. Usage:

import SpatialGSP as GSP
warp_graph_trait = UseSpatialGSP(GSP)

LazyGPR.UseInterpolations — Type

struct UseInterpolations

This is a trait data type that indicates Interpolations.jl should be used to construct a warp function from a set of warp samples. Usage:

import Interpolations
itp_trait = UseInterpolations(Interpolations)

LazyGPR.UseScatteredInterpolation — Type

struct UseScatteredInterpolation

This is a trait data type that indicates ScatteredInterpolation.jl should be used to construct a warp function from a set of warp samples. Usage:

import ScatteredInterpolation
itp_trait = UseScatteredInterpolation(ScatteredInterpolation)

Kernels

LazyGPR.WendlandSplineKernel — Type

struct WendlandSplineKernel{T <: AbstractFloat, N} <: DistanceKernel
    a::T

    # buffer.
    l::Int
    c::NTuple{N,T}
    m::Int
end

The Wendland spline kernels, which are compactly supported. See (Wendland 2004, DOI: 10.1017/CBO9780511617539) for more information.

a controls the bandwidth. Must be larger than zero.

l, c, m are internal intermediate objects that represent a Wendland spline of a specified order and dimension.

LazyGPR.Order1 — Type

struct Order1

This is a trait data type that indicates the order is 1.

LazyGPR.Order2 — Type

struct Order2

This is a trait data type that indicates the order is 2.

LazyGPR.Order3 — Type

struct Order3

This is a trait data type that indicates the order is 3.

LazyGPR.SqExpKernel — Type

struct SqExpKernel{T <: AbstractFloat} <: SqDistanceKernel a::T end

The square exponential kernel. a must be larger than zero. If τ is the distance input, then the kernel output is:

exp(-a*τ)

LazyGPR.DEKernel — Type

struct DEKernel{T <: AbstractFloat, WT, KT <: StationaryKernel} <: DEKernelFamily
    canonical::KT
    warpmap::WT # this is a callable function.
    κ::T
end

The dimensional expansion (DE) kernel container data type.

canonical is the canonical kernel. It should be a stationary kernel.

warpmap is the warp function before the application of the gain multiplier.

κ is the gain. Must be non-negative.

If x, z are inputs, then the kernel evaluates to:

evalkernel(
    canonical,
    norm( [x; κ*warpmap(x)] - [z; κ*warpmap(z)] ),
)

LazyGPR.evalkernel — Function

evalkernel(τ::Real, θ::WendlandSplineKernel{T,N})::T where {T <: AbstractFloat, N}

Evaluates the Wendland spline kernel θ.

evalkernel(τ::Real, θ::ExpKernel{T})::T where T <: AbstractFloat

Evaluates the exponential (Ornstein–Uhlenbeck) kernel θ.

evalkernel(τ::Real, θ::Matern3Halfs{T})::T where T <: AbstractFloat

Evaluates the Matern 3/2 kernel θ.

evalkernel(τ_sq::Real, θ::RQKernel{T,ET})::T where {T <: AbstractFloat, ET}

Evaluates the rational quadratic kernel θ.

evalkernel(τ_sq::Real, θ::SqExpKernel{T})::T where T <: AbstractFloat

Evaluates the square exponential kernel θ.

evalkernel(x::Real, dek::DEKernelFamily)

Evaluates the canonical kernel in dek with the scalar input x. Note that DE kernels can only be evaluated for vector inputs.

LazyGPR.evalde — Function

evalde(dek::DEKernel{T}, x)::T where T <: AbstractFloat

Applies the gain κ.

Evaluates:

dek.κ*dek.warpmap(x)

LazyGPR.evalwarpmap — Function

evalde(dek::DEKernel{T}, x)::T where T <: AbstractFloat

Does not apply the gain κ.

Evaluates:

dek.warpmap(x)

Model specification

LazyGPR.AdjustmentMap — Type

struct AdjustmentMap{T <: AbstractFloat}
    a::T
    b::T
    L::Int
end

Container type for the observation variance adjustment map.

LazyGPR.GPData — Type

struct GPData{T, D, XT <: Union{Vector, SpatialSearchContainer, NTuple}}
    σ²::T
    inputs::XT
    outputs::Array{T,D}
end

Container type for a conventional GPR model.

LazyGPR.create_local_procs — Function

create_local_procs(
    n_new_workers::Integer;
    pkg_load_list::Vector{Expr} = Vector{Expr}(undef, 0),
    verbose = true,
)

Creates n_new_workers many new workers, and loads the packages in pkg_load_list on each worker.

Please manually check whether your current Julia session can actually benefit from launching n_new_workers new workers. For example, the number of workers should not exceed the number of physical floating-point cores in your distributed computing environment.

LazyGPR.evalsmap — Function

evalsmap(x, s::AdjustmentMap{T}) where T

Evaluates the ξ curve from the adjustment map container s, at distance x.

function evalsmap(u, x, s::AdjustmentMap)

Evaluates:

evalsmap(norm(u-x), s)

LazyGPR.LazyGP — Type

struct LazyGP{T, ST <: Union{Nothing, AdjustmentMap{T}}, XT, D}
    b_x::T
    s_map::ST
    data::GPData{T, D, XT}
end

Container type for a the lazy-evaluation GPR model.

b_x is the local neighborhood radius.

s_map is the adjustment map for the observation noise model.

data contains the GPR data.

LazyGPR.fitGP — Function

fitGP(X::Vector, y::Vector{T}, σ²::AbstractFloat, θ) where T <: AbstractFloat

Fit a conventional GPR model.

U: kernel matrix buffer. Mutates.

buffer: pair-wise distance buffer. Mutates.

y: training outputs.

σ²: observation noise variance.

θ: the covariance function.

Returns an object of type DenseGPModel.

LazyGPR.fitGP! — Function

fitGP!(
    U::Matrix{T}, # mutates.
    buffer::RcBuffer, # mutates.
    y::Vector{T},
    σ²::T,
    θ::PositiveDefiniteKernel,
) where T

Fit a conventional GPR model.

U: kernel matrix buffer. Mutates.

buffer: pair-wise distance buffer. Mutates.

y: training outputs.

σ²: observation noise variance.

θ: the covariance function.

Returns:

c: the cached coefficients for querying any predictive mean.

C: the Cholesky factor for U = K + σ²I, where K is the kernel matrix.

Model query

LazyGPR.QueryOptions — Type

@kwdef struct QueryOptions{MT <: BinarySelection, VT <: BinarySelection}
    compute_mean::MT = Enable()
    compute_variance::VT = Enable()
end

Specifies whether the predicitve mean and/or the predictive variance should be computed during querying.

Possible values for each are: LazyGPR.Enabled() and LazyGPR.Disabled().

LazyGPR.computecachevars — Function

computecachevars(::StationaryKernel, args...)

computecachevars(dek::DEKernel, model::LazyGP)

Computes and stores the warpmap evaluations at each training input in model.

computecachevars(dek::DEKernel, data::GPData)

LazyGPR.setup_query_dc — Function

setup_query_dc(
    worker_list,
    model::LazyGP,
    θ::PositiveDefiniteKernel,
    options::QueryOptions,
    cvars::CacheInfo,
)

Loads the neccessary data to the workers in worker_list in preparation to run batch query for the lazy-evaluation GPR model, model, under a distributed computing setup.

LazyGPR.query_dc — Function

query_dc(Xq::Vector, worker_list::Vector{Int})

Query the lazy-evaluation GPR model that was setup by a previous call to setup_query_dc.

The user should call free_query_dc after this function.

query_dc(Xqrs::NTuple{D, AR}, worker_list::Vector{Int}) where {D, AR <: AbstractRange}

LazyGPR.free_query_dc — Function

free_query_dc(worker_list)

Free the data used by the setup_query_dc and query_dc() functions.

LazyGPR.queryGP — Function

queryGP(xrs::NTuple{D, AR}, θ, model::DenseGPModel) where {D, AR <: AbstractRange}

Query a conventional GPR at input xrs.

queryGP(
    xq::AbstractVector{T},
    dek::DEKernel,
    model::DenseGPModel{T},
) where T <: AbstractFloat